[add]動き

+++ Assets/Scripts/Runtime/5.InfraStructure/MinionAsset.cs
@@ -13,13 +13,13 @@ namespace InfraStructure
         /// <summary> ミニオンのタイプ </summary>
         public MinionType MinionType => _minionType;
         /// <summary> 体力 </summary>
-        public int Health => _health;
+        public float Health => _health;
         /// <summary> 防御力 </summary>
-        public int Defense => _defense;
+        public float Defense => _defense;
         /// <summary> 移動速度 </summary>
         public float MoveSpeed => _moveSpeed;
         /// <summary> 攻撃力 </summary>
-        public int AttackPower => _attackPower;
+        public float AttackPower => _attackPower;
         /// <summary> 攻撃距離 </summary>
         public float AttackRange => _attackRange;
         /// <summary> 攻撃速度 </summary>
@@ -35,16 +35,16 @@ namespace InfraStructure
         [SerializeField, Tooltip("ミニオンのタイプ")] private MinionType _minionType;
 
         [Header("体力")]
-        [SerializeField, Tooltip("体力")] private int _health;
+        [SerializeField, Tooltip("体力")] private float _health;
 
         [Header("防御力")]
-        [SerializeField, Tooltip("防御力")] private int _defense;
+        [SerializeField, Tooltip("防御力")] private float _defense;
 
         [Header("移動速度")]
         [SerializeField, Tooltip("移動速度")] private float _moveSpeed;
 
         [Header("攻撃力")]
-        [SerializeField, Tooltip("攻撃力")] private int _attackPower;
+        [SerializeField, Tooltip("攻撃力")] private float _attackPower;
 
         [SerializeField, Tooltip("攻撃距離")] private float _attackRange;
